refactor(wave-4.3): extract shared tiers, tooltip, and date helpers
- Extract toLocalDateStr to activityBuckets.ts; remove local copies from
  streaks.ts and contributionCalendar.ts
- Extract tierColor + computeActivityTiers to lib/usage/chartTiers.ts;
  remove three identical copies from chart components (was computeTiers /
  computeQuantileTiers with same logic in every file)
- Extract ChartTooltip component to ui/ChartTooltip.tsx; replace three
  verbatim tooltip <div> blocks with <ChartTooltip>
- Add ChartLabel local helper in UsageDashboard; replace three identical
  chart-sub-label divs with <ChartLabel>
- Fix: ContributionCalendar month parsing used new Date("YYYY-MM-DD") which
  parses as UTC midnight — wrong month in negative-offset TZs. Now uses
  Number(cell.date.slice(5,7)) - 1 (string parse, no Date object needed)
- Fix: remove unreachable {!hasData && ...} branch inside ContributionCalendar
  return (already guarded by early return above)
- Fix: simplify dead ternary `cells.length > 0 ? weekIndex+1 : 52` — always
  evaluates to weekIndex+1 at this point
