+++ src/components/Heatmap2D.tsx
@@ -2,28 +2,14 @@
 
 import { useState } from "react";
 import type { ActivityBucket } from "@/lib/usage/types";
+import { computeActivityTiers, tierColor } from "@/lib/usage/chartTiers";
+import { ChartTooltip } from "./ui/ChartTooltip";
 
 interface Props {
   byHourDay: ActivityBucket[][];
 }
 
 const DOW_LABELS = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
-
-export function computeQuantileTiers(matrix: ActivityBucket[][]): number[] {
-  const flat = matrix.flat().map((b) => b.turns).filter((v) => v > 0).sort((a, b) => a - b);
-  if (flat.length === 0) return [0, 0, 0, 0, 0];
-  const q = (p: number) => flat[Math.floor(p * (flat.length - 1))];
-  return [q(0.2), q(0.4), q(0.6), q(0.8), flat[flat.length - 1]];
-}
-
-function tierColor(turns: number, tiers: number[]): string {
-  if (turns === 0) return "var(--bg-elevated)";
-  if (turns <= tiers[0]) return "color-mix(in oklch, var(--accent) 30%, transparent)";
-  if (turns <= tiers[1]) return "color-mix(in oklch, var(--accent) 50%, transparent)";
-  if (turns <= tiers[2]) return "color-mix(in oklch, var(--accent) 65%, transparent)";
-  if (turns <= tiers[3]) return "color-mix(in oklch, var(--accent) 80%, transparent)";
-  return "var(--accent)";
-}
 
 export function Heatmap2D({ byHourDay }: Props) {
   const [tooltip, setTooltip] = useState<{ x: number; y: number; content: string } | null>(null);
@@ -37,7 +23,7 @@ export function Heatmap2D({ byHourDay }: Props) {
     );
   }
 
-  const tiers = computeQuantileTiers(byHourDay);
+  const tiers = computeActivityTiers(byHourDay.flat().map((b) => b.turns));
   const cellW = 10;
   const cellH = 10;
   const gap = 2;
@@ -49,7 +35,6 @@ export function Heatmap2D({ byHourDay }: Props) {
   return (
     <div style={{ position: "relative", userSelect: "none" }}>
       <svg width={svgW} height={svgH} style={{ overflow: "visible", width: "100%", height: "auto" }} viewBox={`0 0 ${svgW} ${svgH}`}>
-        {/* Hour labels at 0, 6, 12, 18 */}
         {[0, 6, 12, 18].map((h) => (
           <text
             key={h}
@@ -61,7 +46,6 @@ export function Heatmap2D({ byHourDay }: Props) {
             {h}
           </text>
         ))}
-        {/* Day rows */}
         {byHourDay.map((hourRow, dow) => (
           <g key={dow}>
             <text
@@ -101,27 +85,7 @@ export function Heatmap2D({ byHourDay }: Props) {
           </g>
         ))}
       </svg>
-      {tooltip && (
-        <div
-          style={{
-            position: "fixed",
-            left: tooltip.x,
-            top: tooltip.y,
-            transform: "translate(-50%, -100%)",
-            background: "var(--bg-card)",
-            border: "1px solid var(--border-subtle)",
-            borderRadius: "6px",
-            padding: "5px 9px",
-            fontSize: "11px",
-            color: "var(--text-primary)",
-            pointerEvents: "none",
-            whiteSpace: "nowrap",
-            zIndex: 9999,
-          }}
-        >
-          {tooltip.content}
-        </div>
-      )}
+      {tooltip && <ChartTooltip x={tooltip.x} y={tooltip.y} content={tooltip.content} />}
     </div>
   );
 }
