+++ src/lib/usage/activityBuckets.ts
@@ -2,6 +2,14 @@ import type { ActivityBucket, StreakStats, ContributionCell } from "./types";
 
 export type ActivityTurnInput = { timestamp: string; cost?: number };
 
+export function toLocalDateStr(ts: string): string {
+  const d = new Date(ts);
+  const y = d.getFullYear();
+  const m = String(d.getMonth() + 1).padStart(2, "0");
+  const day = String(d.getDate()).padStart(2, "0");
+  return `${y}-${m}-${day}`;
+}
+
 export interface ActivityData {
   byHourOfDay: ActivityBucket[];
   byDayOfWeek: ActivityBucket[];
